refactor core type mapper to separate classes
