+++ bundles/org.smarthomej.binding.knx/README.md
@@ -6,11 +6,12 @@ Switching lights on and off, activating your roller shutters or changing room te
 To access your KNX bus you either need a gateway device which is connected to the KNX bus and allows computers to access the bus communication.
 This can be either an Ethernet (as a Router or a Tunnel type) or a serial gateway.
 The KNX binding then can communicate directly with this gateway.
-Alternatively a PC running [KNXD](https://github.com/knxd/knxd) (free open source component sofware) can be put in between which then acts as a broker allowing multiple client to connect to the same gateway.
+Alternatively a PC running [KNXD](https://github.com/knxd/knxd) (free open source component software) can be put in between which then acts as a broker allowing multiple client to connect to the same gateway.
 Since the protocol is identical, the KNX binding can also communicate with it transparently.
 
 ***Attention:*** With the introduction of UoM support in version 3.2.7 (see `number` channel below) some data types have changed:
-- the data type for DPT 5.001 (Percent 8bit, 0 -> 100%) has changed from `PercentType` to `QuantityType`for `number` channels (`dimmer`, `color`, `rollershutter` channels stay with `PercentType`)
+
+- the data type for DPT 5.001 (Percent 8bit, 0 -> 100%) has changed from `PercentType` to `QuantityType` for `number` channels (`dimmer`, `color`, `rollershutter` channels stay with `PercentType`)
 - the data type for DPT 5.004 (Percent 8bit, 0 -> 255%) has changed from `PercentType` to `QuantityType`
 - the data type for DPT 6.001 (Percent 8bit -128 -> 127%) has changed from `PercentType` to `QuantityType`
 - the data type for DPT 9.007 (Humidity) has changed from `PercentType` to `QuantityType`
